stuff
do spawners, do items, then do course of game, then main menu, then introduction to intro scene

+++ BirthdayGame/Assets/EnemyFollow.cs
@@ -9,9 +9,9 @@ public class EnemyFollow : MonoBehaviour { //It's more of an "EnemyController".
 
     private LevelManagement GetLevelManagement;
 
-    public int Enemy_Health;
+    public float Enemy_Health;
     public static int PlayerToEnemy_Damage = 1;
-    public int EnemyToPlayer_Damage = 1;
+    public float EnemyToPlayer_Damage = 1;
 
     private void Awake()
     {
